fix matchRoute in RoutingUtils and clean up implementations in XdsNameResolver.
